requirements.txt
altgraph==0.17.4
packaging==23.2
pefile==2023.2.7
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
